backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-bigquery>=3.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0
pillow>=10.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
anthropic>=0.39.0
